Simplify the way serializer metadata is accessed
Dropped class string normalization - I can just raise when registering
the same class name twice.  This means the only map needed is
class string -> class.  From class all of the metadata is accessible.

This may be a temporary change - I believe it's possible to move all
of the data into the class itself, and greatly simplify the format str
so there's just bitstring_format and serializable_format, with the
second being a list of (attr_name, format_field) where format_field is
either a string (raw bitstring format) or a registered class.